feat(transformer/async-to-generator): handle this within arrows function
