+++ crates/oxc_transformer/src/es2017/async_to_generator.rs
@@ -73,6 +73,10 @@ impl<'a, 'ctx> AsyncToGenerator<'a, 'ctx> {
 }
 
 impl<'a, 'ctx> Traverse<'a> for AsyncToGenerator<'a, 'ctx> {
+    fn enter_program(&mut self, _program: &mut Program<'a>, _ctx: &mut TraverseCtx<'a>) {
+        self.ctx.arrow_function_to_expression.enable_async_arrow_function();
+    }
+
     fn exit_expression(&mut self, expr: &mut Expression<'a>, ctx: &mut TraverseCtx<'a>) {
         let new_expr = match expr {
             Expression::AwaitExpression(await_expr) => {
@@ -130,14 +134,9 @@ impl<'a, 'ctx> Traverse<'a> for AsyncToGenerator<'a, 'ctx> {
         }
     }
 
-    fn exit_method_definition(
-        &mut self,
-        node: &mut MethodDefinition<'a>,
-        ctx: &mut TraverseCtx<'a>,
-    ) {
-        let function = &mut node.value;
-        if function.r#async && !function.generator && !function.is_typescript_syntax() {
-            self.executor.transform_function_for_method_definition(function, ctx);
+    fn exit_function(&mut self, func: &mut Function<'a>, ctx: &mut TraverseCtx<'a>) {
+        if func.r#async && ctx.parent().is_method_definition() {
+            self.executor.transform_function_for_method_definition(func, ctx);
         }
     }
 }
